requirements.txt
gpiozero ~= 1.6.2
tinyik ~= 2.2.1
PyYAML ~= 6.0.1
numpy ~= 1.24.3
